backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
PyMySQL>=1.1.0
requests>=2.31.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pydantic>=2.6.0,<3.0
httpx>=0.27.0,<0.29.0
python-dotenv>=1.0.0,<2.0
